Run rename-project.bash
Now the repo has been generated we need to update the naming from the template and as a result running the rename-project bash repo and passing in hmpps-integration-events should update the codebase appropriatley
