+++ .circleci/config.yml-z
@@ -6,10 +6,10 @@ orbs:
 parameters:
   alerts-slack-channel:
     type: string
-    default: 
+    default: PIPELINE_SECURITY_SLACK_CHANNEL
   releases-slack-channel:
     type: string
-    default: 
+    default: SLACK_RELEASES_CHANNEL
 
 jobs:
   validate:
