+++ applicationinsights.json
@@ -1,6 +1,6 @@
 {
   "role": {
-    "name": "hmpps-template-kotlin"
+    "name": "hmpps-integration-events"
   },
   "customDimensions": {
     "service.version": "${BUILD_NUMBER}"
