+++ docker-compose.yml
@@ -1,10 +1,10 @@
 version: "3"
 services:
-  hmpps-template-kotlin:
+  hmpps-integration-events:
     build:
       context: .
     network_mode: "host"
-    container_name: hmpps-template-kotlin
+    container_name: hmpps-integration-events
     ports:
       - "8081:8080"
     healthcheck:
