+++ helm_deploy/hmpps-integration-events/Chart.yaml
@@ -1,7 +1,7 @@
 apiVersion: v2
 appVersion: '1.0'
 description: A Helm chart for Kubernetes
-name: hmpps-template-kotlin
+name: hmpps-integration-events
 version: 0.2.0
 dependencies:
   - name: generic-service
