+++ helm_deploy/values-preprod.yaml-z
@@ -13,4 +13,4 @@ generic-service:
 # CloudPlatform AlertManager receiver to route prometheus alerts to slack
 # See https://user-guide.cloud-platform.service.justice.gov.uk/documentation/monitoring-an-app/how-to-create-alarms.html#creating-your-own-custom-alerts
 generic-prometheus-alerts:
-  alertSeverity: 
+  alertSeverity: NON_PROD_ALERTS_SEVERITY_LABEL
