+++ helm_deploy/values-prod.yaml
@@ -1,11 +1,11 @@
 ---
-# Per environment values which override defaults in hmpps-template-kotlin/values.yaml
+# Per environment values which override defaults in hmpps-integration-events/values.yaml
 
 generic-service:
   ingress:
-    host: hmpps-template-kotlin.hmpps.service.justice.gov.uk
+    host: hmpps-integration-events.hmpps.service.justice.gov.uk
 
 # CloudPlatform AlertManager receiver to route prometheus alerts to slack
 # See https://user-guide.cloud-platform.service.justice.gov.uk/documentation/monitoring-an-app/how-to-create-alarms.html#creating-your-own-custom-alerts
 generic-prometheus-alerts:
-  alertSeverity: PROD_ALERTS_SEVERITY_LABEL
+  alertSeverity:
