+++ helm_deploy/values-prod.yaml-z
@@ -2,15 +2,10 @@
 # Per environment values which override defaults in hmpps-integration-events/values.yaml
 
 generic-service:
-  replicaCount: 2
-
   ingress:
-    host: hmpps-integration-events-preprod.hmpps.service.justice.gov.uk
-
-  env:
-    APPLICATIONINSIGHTS_CONFIGURATION_FILE: applicationinsights.dev.json
+    host: hmpps-integration-events.hmpps.service.justice.gov.uk
 
 # CloudPlatform AlertManager receiver to route prometheus alerts to slack
 # See https://user-guide.cloud-platform.service.justice.gov.uk/documentation/monitoring-an-app/how-to-create-alarms.html#creating-your-own-custom-alerts
 generic-prometheus-alerts:
-  alertSeverity: 
+  alertSeverity: PROD_ALERTS_SEVERITY_LABEL
