+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationevents/health/HealthInfo.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppstemplatepackagename.health
+package uk.gov.justice.digital.hmpps.hmppsintegrationevents.health
 
 import org.springframework.boot.actuate.health.Health
 import org.springframework.boot.actuate.health.HealthIndicator
