+++ src/main/resources/application.yml
@@ -1,10 +1,10 @@
 info.app:
-  name: HMPPS Template Kotlin
+  name: Hmpps Integration Events
   version: 1.0
 
 spring:
   application:
-    name: hmpps-template-kotlin
+    name: hmpps-integration-events
   codec:
     max-in-memory-size: 10MB
