+++ src/main/resources/logback-spring.xml
@@ -12,7 +12,7 @@
     </encoder>
   </appender>
 
-  <logger name="uk.gov.justice.digital.hmpps.hmppstemplatepackagename.HmppsTemplateKotlinKt" additivity="false" level="DEBUG">
+  <logger name="uk.gov.justice.digital.hmpps.hmppsintegrationevents.HmppsIntegrationEventsKt" additivity="false" level="DEBUG">
     <appender-ref ref="consoleAppender"/>
   </logger>
