+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationevents/health/HealthInfoTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppstemplatepackagename.health
+package uk.gov.justice.digital.hmpps.hmppsintegrationevents.health
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
