+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationevents/integration/NotFoundTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppstemplatepackagename.integration
+package uk.gov.justice.digital.hmpps.hmppsintegrationevents.integration
 
 import org.junit.jupiter.api.Test
